Add force_resize_to_same_host option
Instance resize is kind of migration, but not block-migration, so it cannot be
performed without shared storage for instance drives if resized instance is
scheduled to other node (not the node it ran on before resize).

This commit adds option force_resize_to_same_host that forces resized instance
to be scheduled to the node it was running before resize, which fixes
the problem.

Issue: altai-150
Change-Id: Id605b8fa0c3181344cd71c8790e8cbc65785a85d